Splitting of KafkaPublish Tests
